requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
confuse==1.1.0
Deprecated==1.2.9
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyGithub==1.50
PyJWT==1.7.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
wrapt==1.12.1
